@@ -10,36 +10,39 @@
     "prepare": "husky install"
   },
   "dependencies": {
-    "@headlessui/react": "^1.7.15",
+    "@headlessui/react": "^2.2.0",
     "@monaco-editor/react": "^4.5.1",
     "@tailwindcss/forms": "^0.5.4",
-    "@types/node": "20.3.2",
-    "@types/react": "18.2.14",
-    "@types/react-dom": "18.2.6",
     "allotment": "^1.19.0",
-    "autoprefixer": "10.4.14",
-    "eslint": "8.43.0",
-    "eslint-config-next": "13.4.7",
-    "html-to-image": "^1.11.11",
+    "html-to-image": "1.11.11",
     "jsonc-parser": "^3.2.0",
     "lodash.debounce": "^4.0.8",
-    "lodash.get": "^4.4.2",
     "lodash.set": "^4.3.2",
-    "next": "13.4.7",
-    "postcss": "8.4.24",
-    "prettier": "^2.8.8",
-    "prettier-plugin-tailwindcss": "^0.3.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
     "react-zoom-pan-pinch": "^3.1.0",
-    "reaflow": "5.1.2",
-    "tailwindcss": "3.3.2",
-    "typescript": "5.1.6",
-    "zustand": "^4.3.9"
+    "reaflow": "^5.3.4",
+    "zustand": "^5.0.3"
   },
   "devDependencies": {
+    "@eslint/eslintrc": "^3.3.1",
+    "@eslint/js": "^9.23.0",
+    "@next/eslint-plugin-next": "^15.2.4",
+    "@tailwindcss/postcss": "^4.0.17",
     "@types/lodash.debounce": "^4.0.7",
-    "husky": "^8.0.3",
-    "lint-staged": "^13.2.3"
+    "@types/node": "^22.13.14",
+    "@types/react": "^19.0.12",
+    "@types/react-dom": "^19.0.4",
+    "eslint": "^9.23.0",
+    "eslint-config-next": "^15.2.4",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "husky": "^9.1.7",
+    "lint-staged": "^15.5.0",
+    "next": "^15.2.4",
+    "postcss": "^8.5.3",
+    "prettier": "^3.5.3",
+    "prettier-plugin-tailwindcss": "^0.6.11",
+    "tailwindcss": "4.0.17",
+    "typescript": "^5.8.2"
   }
 }
