@@ -1,9 +1,9 @@
-module.exports = {
+export default {
   bracketSpacing: true,
   semi: true,
   singleQuote: false,
   trailingComma: "all",
   printWidth: 80,
   tabWidth: 2,
-  plugins: [require("prettier-plugin-tailwindcss")],
+  plugins: [await import("prettier-plugin-tailwindcss")],
 };
